Create Pi_Gui.py
This is to be downloaded on the pi for a final iteration of the GUI. I made it a separate file for easy distinguishing. THIS WILL NOT STREAM VIDEO. It will however: display the GUI on a scaled ratio fitted to your screen, control a servo through the GPIO pins, and (if neccessary) quit the program all together. 
Still more cleaning up to do though.